+++ package/src/index.js
@@ -10,8 +10,10 @@ export * from './input';
 export * from './link';
 export * from './list';
 export * from './locale-context';
+export * from './menu';
 export * from './modal';
 export * from './pdf-viewer';
+export * from './popover';
 export * from './root';
 export * from './row';
 export * from './rs-input';
